eager load tags - much faster
https://github.com/mbleigh/acts-as-taggable-on/issues/91

See this issue. Before this, every time I made a call to .tag_list
it was making a call to the database which was making everything take
AGES to load.

After this commit, it will eager load the tags in the jobs controller,
and when searching tags it can then avoid using the tag_list so there's
no extra load time