Store search location on jobs
After this commit, you can see on both a Job and
a ScrapedJob what the location we searched when we
found them.

For example, if we're searching London jobs then the
searched_location will be London even though the
location on the advert might be canary wharf.

This is so we can easily filter all jobs by location.
For example I can search for all London jobs without
having to convert the location they have on the advert
to the areas I originally search for